+++ Config/config.php
@@ -3,7 +3,4 @@
 return [
     'name' => 'ItkIssueCreate',
     'teamsWebHook' => env('TEAMS_WEB_HOOK', null),
-    'leantimeUrl' => env('LEANTIME_URL', null),
-    'leantimeApiKey' => env('LEANTIME_API_KEY', null),
-    'leantimeProjectKey' => env('LEANTIME_PROJECT_KEY', null),
 ];