+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add prometheus logging to post requests
+### Changed
+
+- Removed leantime helper and related config.
 
 ## [1.0.0]
+
+- Add prometheus logging to post requests